move updator to separate module/file
