+++ src/test/definition.rs
@@ -1,7 +1,7 @@
+use crate::test;
 use crate::test::file;
 use crate::test::http;
 use crate::test::validation;
-use crate::test;
 use serde::{Deserialize, Serialize};
 use std::cell::Cell;
